update branch according to main, implement issue changes
