+++ src/components/SingleList.jsx
@@ -1,8 +1,12 @@
 import './SingleList.css';
+import { useNavigate } from 'react-router-dom';
 
 export function SingleList({ name, path, setListPath }) {
+	const navigate = useNavigate();
+
 	function handleClick() {
 		setListPath(path);
+		navigate('/list');
 	}
 
 	return (
